Initial check-in of (not working) code
Ugh, I should have checked this in when it was incomplete instead of
breaking it and then checking it in half-way before reimplementing
status change handling.

Right now it's pretty broken.

It used to keep the various state changes explicitly in specific fields,
but in the half-finished reimplementation they're stored as Status
Changes.

More general, somewhat more complicated.

It also currently uses SQLite whereas the eventual tool should use
MySQL, but *hopefully* that should be a relatively trivial change
because of Gorm's abstraction.