Fixes list options, adds examples text
So, in an earlier change I totally broke the different kinds of list,
and also made actually determining the status of any given exception
really difficult in SQL.

So, now status is double-tracked: both by an actual status property, and
by the audit log of changes to the status.

This is much easier at the slight risk of being able to become
inconsistent.